Fix build: bo weakSelf trung trong viewDidLoad (menu Cong cu da khai bao) + cast bitmapInfo
Daemon (JS engine) bien dich OK; loi o app: TVNCRootListController.m khai
bao weakSelf hai lan trong viewDidLoad (dong them cho menu + dong cu cho
nw_path_monitor). Bo cai trung. Cast CGBitmapContextCreate flags cho het
warning enum OR.

+++ src/trollvncserver.mm
@@ -4667,8 +4667,9 @@ static uint8_t *tvLoadImageRGBA(NSString *path, int *outW, int *outH) {
     if (!buf)
         return NULL;
     CGColorSpaceRef cs = CGColorSpaceCreateDeviceRGB();
-    CGContextRef c = CGBitmapContextCreate(buf, w, h, 8, w * 4, cs,
-                                           kCGImageAlphaPremultipliedLast | kCGBitmapByteOrder32Big);
+    CGContextRef c = CGBitmapContextCreate(
+        buf, w, h, 8, w * 4, cs,
+        (uint32_t)(kCGImageAlphaPremultipliedLast | kCGBitmapByteOrder32Big));
     CGColorSpaceRelease(cs);
     if (!c) {
         free(buf);
